fixed hidden empty line leak, a bit of norm and almost finished making cat play well with other cats

+++ srcs/exec.c
@@ -41,6 +41,35 @@ static void	child_proces(int *p, t_cmd *cmd)
 	close(p[0]);
 }
 
+int	loop_part1(t_cmd **cmd, char **path)
+{
+	if (check_if_builtin_not_pipe(*cmd))
+	{
+		_shell()->exit_code = run_builtin_not_piped(*cmd);
+		if (!((*cmd)->next) && !((*cmd)->prev) && ft_strexact("exit",
+				(*cmd)->args[0]))
+			_shell()->exit = 1;
+		*cmd = (*cmd)->next;
+		return (1);
+	}
+	*path = NULL;
+	if (!check_if_builtin_2pipe(*cmd))
+		*path = ft_get_exec_path((*cmd)->args);
+	if (!check_if_builtin_2pipe(*cmd) && !(*path))
+	{
+		printf("bash: %s: command not found\n", *((*cmd)->args));
+		_shell()->exit_code = 127;
+		*cmd = (*cmd)->next;
+		return (1);
+	}
+	if ((*cmd)->flag && (*cmd)->next && (ft_strexact("ls",
+				((*cmd)->next)->args[0]) || ft_strexact("pwd",
+				((*cmd)->next)->args[0]) || (ft_strexact("cat",
+					((*cmd)->next)->args[0]))))
+		(*cmd)->flag = 2;
+	return (0);
+}
+
 void	loop_execution(t_cmd *cmd)
 {
 	int		p[2];
@@ -55,34 +84,10 @@ void	loop_execution(t_cmd *cmd)
 	bs_cat = 0;
 	while (cmd)
 	{
-		if (*cmd->args == NULL)
-		{
-			cmd = cmd->next;
+		if (loop_part1(&cmd, &path))
 			continue ;
-		}
-		if (check_if_builtin_not_pipe(cmd))
-		{
-			_shell()->exit_code = run_builtin_not_piped(cmd);
-			if (!cmd->next && !cmd->prev && ft_strexact("exit", cmd->args[0]))
-				_shell()->exit = 1;
-			cmd = cmd->next;
-			continue ;
-		}
-		path = NULL;
-		if (!check_if_builtin_2pipe(cmd))
-			path = ft_get_exec_path(cmd->args);
-		if (!check_if_builtin_2pipe(cmd) && !path)
-		{
-			printf("bash: %s: command not found\n", *(cmd->args));
-			_shell()->exit_code = 127;
-			cmd = cmd->next;
-			continue ;
-		}
 		if (cmd)
 			cmd->fd_in = p[0];
-		if (cmd->flag && cmd->next && (ft_strexact("ls", (cmd->next)->args[0])
-				|| ft_strexact("pwd", (cmd->next)->args[0])))
-			cmd->flag = 2;
 		pipe(p);
 		pid = fork();
 		if (pid == 0)
@@ -96,11 +101,8 @@ void	loop_execution(t_cmd *cmd)
 		}
 		else
 		{
-			if (cmd->flag == 2)
-			{
-				bs_cat++;
+			if (cmd->flag == 2 && ++bs_cat)
 				kill(pid, SIGKILL);
-			}
 			else
 				waitpid(pid, &status, WUNTRACED);
 			close(p[1]);
@@ -116,8 +118,6 @@ void	loop_execution(t_cmd *cmd)
 				}
 			}
 			cmd = cmd->next;
-			// if (cmd)
-			// 	cmd->fd_in = p[0];
 			if (path)
 				free(path);
 		}
